redis/todoproject/requirements.txt
aiofiles==22.1.0
asgiref==3.6.0
async-timeout==4.0.2
click==8.1.3
colorama==0.4.6
corn==8
dict-toolbox==2.3.1
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.4
django-filter==22.1
django-redis==5.2.0
django-rq==2.6.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
environs==9.5.0
filelock==3.8.2
Jinja2==3.1.2
lazy-object-proxy==1.8.0
Markdown==3.4.1
MarkupSafe==2.1.1
marshmallow==3.19.0
msgpack==1.0.4
nest-asyncio==1.5.6
packaging==22.0
pbr==5.11.0
platformdirs==2.6.0
pop==23.0.0
pop-config==10.2.3
pop-loop==1.0.6
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.7
PyYAML==6.0
redis==4.4.0
rend==6.5.1
rq==1.11.1
sniffio==1.3.0
sqlparse==0.4.3
stevedore==4.1.1
toml==0.10.2
virtualenv==20.17.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
